total refactoring: protocol implemented by formats; recv entire buffer  and parsing after; maze generator library adapted for usage in game

+++ headers/list.h
@@ -13,7 +13,7 @@ struct listElements_t
 {
     struct element_t*   first;
     struct element_t*   last;
-    uint8_t             count;
+    size_t              count;
 };
